refactor: split index into components
