+++ app/components/conference/Past.vue
@@ -1,0 +1,26 @@
+<template>
+  <LayoutContainer>
+    <LayoutContainerContent>
+      <UiHeading class="mb-12 text-vue">
+        {{ $t('headings.look_back_on_2024') }}
+      </UiHeading>
+
+      <i18n-t
+        tag="p"
+        scope="global"
+        keypath="pages.home.look_back_on_2024"
+        class="mb-12 text-base max-w-[128ch] text-center text-balance"
+      />
+
+      <NuxtLink to="/2024">
+        <UiImage
+          class="w-auto aspect-[1200/971] max-h-[330px] mx-auto"
+          fit="contain"
+          src="/pragvue2024.webp"
+          :title="$t('pages.home.link_2024')"
+          :alt="$t('accessibility.photo_of_the_speakers_from_2024')"
+        />
+      </NuxtLink>
+    </LayoutContainerContent>
+  </LayoutContainer>
+</template>
